Moved database metadata into database/metadata
