+++ app/assets/js/src/database/adapterV1/getDayTasksV1.ts
@@ -1,9 +1,9 @@
 import type { DayTaskInfo } from 'data/dayTasks';
 import type { DatabaseData } from 'data/shared/types';
-import { ObjectStoreName } from 'utils';
 import { getDatabase, getIdbRequestPromise } from 'utils/indexedDB';
 import { getStatuses } from '../getStatuses';
 import type { LegacyStatusName } from 'data/shared/types/LegacyExportDataVersions';
+import { ObjectStoreName } from 'database/metadata';
 
 /**
  * Retrieve all schema v1 {@linkcode DayTaskInfo} information from the database v2.
