+++ app/assets/js/src/database/metadata/index.ts
@@ -1,0 +1,5 @@
+export { ObjectStoreName } from './ObjectStoreName';
+export { IndexName } from './IndexName';
+
+export const dbName = 'orange-twist';
+export const dbVersion = 2;
