+++ app/assets/js/src/database/utils/getDatabaseVersion.ts
@@ -1,10 +1,12 @@
+import { dbName } from 'database/metadata';
+
 /**
  * Determine the version of the existing database, or return `null` if the database doesn't exist yet.
  */
 export async function getDatabaseVersion(): Promise<number | null> {
 	const existingDbs = await indexedDB.databases();
 	// TODO: Avoid this magic string
-	const existingDb = existingDbs.find(({ name }) => name === 'orange-twist');
+	const existingDb = existingDbs.find(({ name }) => name === dbName);
 	const existingDbVersion = existingDb?.version;
 	return existingDbVersion ?? null;
 }
