+++ app/assets/js/src/database/utils/index.ts
@@ -1,0 +1,1 @@
+export { getDatabaseVersion } from './getDatabaseVersion';
