+++ app/assets/js/src/utils/indexedDB/index.ts
@@ -1,7 +1,5 @@
 export { doDatabaseTransaction } from './doDatabaseTransaction';
 export { getDatabase } from './getDatabase';
-export { getDatabaseVersion } from './getDatabaseVersion';
 export { getDbDump } from './getDbDump';
 export { getEntries } from './getEntries';
 export { getIdbRequestPromise } from './getIdbRequestPromise';
-export { ObjectStoreName } from '../../data/shared/ObjectStoreName';
